Clarify handler and router Relates #19
