+++ src/server.js
@@ -1,12 +1,14 @@
 const http = require("http");
 const fs = require("fs");
 const path = require("path");
-const functions = require("./handler.js");
+const router = require("./router.js");
 
 const port = 3000;
 
-const server = http.createServer(functions.handler);
+const server = http.createServer(router);
 
 server.listen(port, function() {
-	console.log(`Server is listening on port http://localhost:${port} . Ready to accept requests!`);
+    console.log(
+        `Server is listening on port http://localhost:${port} . Ready to accept requests!`
+    );
 });
